+++ tilings/tilescope.py
@@ -1,3 +1,6 @@
+import itertools
+import math
+from array import array
 from collections import Counter
 from typing import Counter as CounterType
 from typing import (
@@ -9,6 +12,7 @@ from typing import (
     Optional,
     Set,
     Tuple,
+    Type,
     Union,
     cast,
 )
@@ -21,14 +25,21 @@ from comb_spec_searcher import (
     CombinatorialSpecification,
     CombinatorialSpecificationSearcher,
 )
+from comb_spec_searcher.class_db import ClassDB, ClassKey, Info, Key
 from comb_spec_searcher.class_queue import CSSQueue, DefaultQueue, WorkPacket
 from comb_spec_searcher.rule_db.abstract import RuleDBAbstract
+from comb_spec_searcher.strategies.rule import AbstractRule
 from comb_spec_searcher.typing import CombinatorialClassType, CSSstrategy
 from permuta import Basis, Perm
 from tilings import GriddedPerm, Tiling
+from tilings.assumptions import TrackingAssumption
 from tilings.strategy_pack import TileScopePack
 
 __all__ = ("TileScope", "TileScopePack", "LimitedAssumptionTileScope", "GuidedSearcher")
+
+Cell = Tuple[int, int]
+TrackedClassAssumption = Tuple[int, Tuple[Cell, ...]]
+TrackedClassDBKey = Tuple[int, Tuple[TrackedClassAssumption, ...]]
 
 
 class TileScope(CombinatorialSpecificationSearcher):
@@ -42,6 +53,8 @@ class TileScope(CombinatorialSpecificationSearcher):
         start_class: Union[str, Iterable[Perm], Tiling],
         strategy_pack: TileScopePack,
         ruledb: Optional[RuleDBAbstract] = None,
+        classdb: Optional[ClassDB] = None,
+        classqueue: Optional[CSSQueue] = None,
         expand_verified: bool = False,
         debug: bool = False,
     ) -> None:
@@ -74,7 +87,9 @@ class TileScope(CombinatorialSpecificationSearcher):
         super().__init__(
             start_class=start_tiling,
             strategy_pack=strategy_pack,
+            classdb=classdb,
             ruledb=ruledb,
+            classqueue=classqueue,
             expand_verified=expand_verified,
             debug=debug,
         )
@@ -94,23 +109,23 @@ class LimitedAssumptionTileScope(TileScope):
         ignore_full_tiling_assumptions: bool = False,
         **kwargs,
     ) -> None:
-        super().__init__(start_class, strategy_pack, **kwargs)
         self.max_assumptions = max_assumptions
+        super().__init__(
+            start_class,
+            strategy_pack,
+            classdb=TrackedClassDB(),
+            **kwargs,
+        )
         self.ignore_full_tiling_assumptions = ignore_full_tiling_assumptions
 
-    def _expand(
-        self,
-        comb_class: CombinatorialClassType,
-        label: int,
-        strategies: Tuple[CSSstrategy, ...],
-        inferral: bool,
-    ) -> None:
+    def _rules_from_strategy(  # type: ignore
+        self, comb_class: CombinatorialClassType, strategy: CSSstrategy
+    ) -> Iterator[AbstractRule]:
         """
-        Will expand the combinatorial class with given label using the given
-        strategies, but only add rules whose children all satisfy the max_assumptions
-        requirement.
+        Yield all the rules given by a strategy/strategy factory whose children all
+        satisfy the max_assumptions constraint.
         """
-
+        # pylint: disable=arguments-differ
         def num_child_assumptions(child: Tiling) -> int:
             return sum(
                 1
@@ -119,18 +134,12 @@ class LimitedAssumptionTileScope(TileScope):
                 or len(ass.gps) != len(child.active_cells)
             )
 
-        if inferral:
-            self._inferral_expand(comb_class, label, strategies)
-        else:
-            for strategy_generator in strategies:
-                for start_label, end_labels, rule in self._expand_class_with_strategy(
-                    comb_class, strategy_generator, label
-                ):
-                    if all(
-                        num_child_assumptions(child) <= self.max_assumptions
-                        for child in rule.children
-                    ):
-                        self.add_rule(start_label, end_labels, rule)
+        for rule in super()._rules_from_strategy(comb_class, strategy):
+            if all(
+                num_child_assumptions(child) <= self.max_assumptions
+                for child in rule.children
+            ):
+                yield rule
 
 
 class GuidedSearcher(TileScope):
@@ -139,11 +148,15 @@ class GuidedSearcher(TileScope):
         tilings: Iterable[Tiling],
         basis: Tiling,
         pack: TileScopePack,
-        *args,
         **kwargs,
     ):
         self.tilings = frozenset(t.remove_assumptions() for t in tilings)
-        super().__init__(basis, pack, *args, **kwargs)
+        super().__init__(
+            basis,
+            pack,
+            classdb=TrackedClassDB(),
+            **kwargs,
+        )
         for t in self.tilings:
             class_label = self.classdb.get_label(t)
             is_empty = self.classdb.is_empty(t, class_label)
@@ -212,14 +225,14 @@ class TrackedSearcher(LimitedAssumptionTileScope):
         **kwargs,
     ) -> None:
         super().__init__(
-            start_class, strategy_pack, max_assumptions=max_assumptions, **kwargs
+            start_class,
+            strategy_pack,
+            max_assumptions=max_assumptions,
+            classqueue=TrackedQueue(
+                cast(TileScopePack, strategy_pack), self, delay_next
+            ),
+            **kwargs,
         )
-        # reset to the trackedqueue!
-        self.classqueue = cast(
-            DefaultQueue,
-            TrackedQueue(cast(TileScopePack, self.strategy_pack), self, delay_next),
-        )  # TODO: make CSS accept a CSSQueue as a kwarg
-        self.classqueue.add(self.start_label)
 
 
 class TrackedDefaultQueue(DefaultQueue):
@@ -388,3 +401,240 @@ class TrackedQueue(CSSQueue):
             except StopIteration:
                 continue
         raise StopIteration("No elements in queue")
+
+
+class TrackedClassDB(ClassDB[Tiling]):
+    def __init__(self) -> None:
+        super().__init__(Tiling)
+        self.classdb = ClassDB(Tiling)
+        self.label_to_tilings: List[bytes] = []
+        self.tilings_to_label: Dict[bytes, int] = {}
+        self.assumption_type_to_int: Dict[Type[TrackingAssumption], int] = {}
+        self.int_to_assumption_type: List[Type[TrackingAssumption]] = []
+
+    def __iter__(self) -> Iterator[int]:
+        for key in self.label_to_info:
+            yield key
+
+    def __contains__(self, key: Key) -> bool:
+        if isinstance(key, Tiling):
+            actual_key = self.tiling_to_key(key)
+            compressed_key = self._compress_key(actual_key)
+            return self.tilings_to_label.get(compressed_key) is not None
+        if isinstance(key, int):
+            return 0 <= key < len(self.label_to_tilings)
+        raise ValueError("Invalid key")
+
+    def __eq__(self, other: object) -> bool:
+        if not isinstance(other, TrackedClassDB):
+            return NotImplemented
+        return bool(
+            self.classdb == other.classdb
+            and self.label_to_tilings == other.label_to_tilings
+            and self.tilings_to_label == other.tilings_to_label
+        )
+
+    def tiling_to_key(self, tiling: Tiling) -> TrackedClassDBKey:
+        """
+        Converts a tiling to its corresponding key.
+        """
+        underlying_label = self.classdb.get_label(tiling.remove_assumptions())
+        assumption_keys = tuple(
+            self.assumption_to_key(ass) for ass in tiling.assumptions
+        )
+        return (underlying_label, assumption_keys)
+
+    def assumption_to_key(self, ass: TrackingAssumption) -> TrackedClassAssumption:
+        """
+        Determines the type of the assumption and retrieves the int representing
+        that type from the appropriate class variables, and then apprends the cells.
+        """
+        try:
+            ass_type_int = self.assumption_type_to_int[type(ass)]
+        except KeyError:
+            ass_type_int = len(self.int_to_assumption_type)
+            assert ass_type_int < 256
+            self.int_to_assumption_type.append(type(ass))
+            self.assumption_type_to_int[type(ass)] = ass_type_int
+        return (ass_type_int, ass.get_cells())
+
+    def key_to_tiling(self, key: TrackedClassDBKey) -> Tiling:
+        """
+        Converts a key back to a Tiling.
+        """
+        return self.classdb.get_class(key[0]).add_assumptions(
+            (
+                self.int_to_assumption_type[ass_key[0]].from_cells(ass_key[1])
+                for ass_key in key[1]
+            ),
+            clean=False,
+        )
+
+    @staticmethod
+    def _compress_key(key: TrackedClassDBKey) -> bytes:
+        # Assumes there are fewer than 256 assumptions
+        # Assumes every assumption covers fewer than 256 cells
+        # Assumes the positions in an assumption have value < 256
+
+        def int_to_bytes(n: int) -> List[int]:
+            """
+            Converts an int to a list of ints all in [0 .. 255] ready for
+            byte compression. First entry is the number of bytes needed (assumes < 256),
+            remaining entries the bytes composing the int from lowest byte up to largest
+            byte.
+            """
+            bytes_needed = max(math.ceil(n.bit_length() / 8), 1)
+            result: List[int] = [bytes_needed]
+            while n >= 2**8:
+                result.append(n & 0xFF)
+                n = n >> 8
+            result.append(n)
+            return result
+
+        def _compress_assumption(ass_key: TrackedClassAssumption) -> List[int]:
+            type_int, cells = ass_key
+            assert type_int < 256
+            assert len(cells) < 256
+            assert all(cell[0] < 256 and cell[1] < 256 for cell in cells)
+
+            result = [type_int]
+            result.append(len(cells))
+            result.extend(itertools.chain(*cells))
+            return result
+
+        result: List[int] = int_to_bytes(key[0])
+        result.extend(
+            itertools.chain.from_iterable(
+                _compress_assumption(ass_key) for ass_key in key[1]
+            )
+        )
+        compressed_key = array("B", result).tobytes()
+        return compressed_key
+
+    @staticmethod
+    def _decompress_key(compressed_key: bytes) -> TrackedClassDBKey:
+        def int_from_bytes(n: array) -> int:
+            """
+            Converts a list of ints to a single int assuming the first entry is the
+            lowest byte and so on.
+            """
+            result = n[0]
+            for idx in range(1, len(n)):
+                result |= n[idx] << (8 * idx)
+            return cast(int, result)
+
+        def _decompress_tuple_of_cells(
+            compressed_cells: array,
+        ) -> Tuple[Cell, ...]:
+            """
+            compressed_cells is a list of 2*i bytes, each of which is a coordinates
+            """
+            vals = iter(compressed_cells)
+            return tuple(
+                (next(vals), next(vals)) for _ in range(len(compressed_cells) // 2)
+            )
+
+        vals = array("B", compressed_key)
+        offset = 0
+
+        num_bytes_int = vals[offset]
+        offset += 1
+        label = int_from_bytes(vals[offset : offset + num_bytes_int])
+        offset += num_bytes_int
+
+        tuples_of_cells = []
+        while offset < len(vals):
+            type_int, num_cells = vals[offset : offset + 2]
+            offset += 2
+            tuples_of_cells.append(
+                (
+                    type_int,
+                    _decompress_tuple_of_cells(vals[offset : offset + 2 * num_cells]),
+                )
+            )
+            offset += 2 * num_cells
+
+        return (label, tuple(tuples_of_cells))
+
+    def add(self, comb_class: ClassKey, compressed: bool = False) -> None:
+        """
+        Adds a Tiling to the classdb
+        """
+        if compressed:
+            raise NotImplementedError
+        if isinstance(comb_class, Tiling):
+            key = self.tiling_to_key(comb_class)
+            compressed_key = self._compress_key(key)
+            if compressed_key not in self.tilings_to_label:
+                self.label_to_tilings.append(compressed_key)
+                self.tilings_to_label[compressed_key] = len(self.tilings_to_label)
+
+    def _get_info(self, key: Key) -> Info:
+        """
+        Return the "Info" object corresponding to the key, which is
+        either a Tiling or an integer
+        """
+        # pylint: disable=protected-access
+        if isinstance(key, Tiling):
+            actual_key = self.tiling_to_key(key)
+            compressed_key = self._compress_key(actual_key)
+            if compressed_key not in self.tilings_to_label:
+                self.add(key)
+            info: Optional[Info] = self.classdb._get_info(actual_key[0])
+            if info is None:
+                raise ValueError("Invalid key")
+            info = Info(
+                key,
+                self.tilings_to_label[compressed_key],
+                info.empty,
+            )
+        elif isinstance(key, int):
+            if not 0 <= key < len(self.label_to_tilings):
+                raise KeyError("Key not in ClassDB")
+            tiling_key = self._decompress_key(self.label_to_tilings[key])
+            info = self.classdb.label_to_info.get(tiling_key[0])
+            if info is None:
+                raise ValueError("Invalid key")
+            info = Info(
+                self.key_to_tiling(tiling_key),
+                key,
+                info.empty,
+            )
+        else:
+            raise TypeError()
+        return info
+
+    def get_class(self, key: Key) -> Tiling:
+        """
+        Return combinatorial class of key.
+        """
+        info = self._get_info(key)
+        return cast(Tiling, info.comb_class)
+
+    def is_empty(self, comb_class: Tiling, label: Optional[int] = None) -> bool:
+        """
+        Return True if combinatorial class is set to be empty, False if not.
+        """
+        return bool(self.classdb.is_empty(comb_class.remove_assumptions()))
+
+    def set_empty(self, key: Key, empty: bool = True) -> None:
+        """
+        Set a class to be empty.
+        """
+        if isinstance(key, int):
+            if 0 <= key < len(self.label_to_tilings):
+                underlying_label, _ = self._decompress_key(self.label_to_tilings[key])
+        if isinstance(key, Tiling):
+            underlying_label = self.classdb.get_label(key.remove_assumptions())
+        self.classdb.set_empty(underlying_label, empty)
+
+    def status(self) -> str:
+        """
+        Return a string with the current status of the run.
+        """
+        status = self.classdb.status()
+        status = status.replace("combinatorial classes", "underlying tilings")
+        tilings = "\n\tTotal number of tilings found is"
+        tilings += f" {len(self.label_to_tilings):,d}"
+        status = status.replace("ClassDB status:", "TrackedClassDB status:" + tilings)
+        return status + "\n"